bidder-main/requirements.txt
requests==2.32.3
schedule==1.2.2
pandas==2.2.3
numpy==2.2.6
psycopg2-binary==2.9.10
statsmodels==0.14.4
